task 1

+++ modules/RemoveDB.ts
@@ -1,10 +1,11 @@
 import ListDBSync from './ListDBSync'
 import fs from 'fs'
 
-function RemoveDB (id: string): boolean {
+async function RemoveDB (id: string): Promise<boolean> {
+  if (ListDBSync.getSyncMode() === 'client') return false
   if (id === undefined) return false
   if (!(ListDBSync.getAllDataId()).includes(id)) return false
-  ListDBSync.removeData(id)
+  await Promise.resolve(ListDBSync.removeData(id))
   fs.rmSync(`${ListDBSync.getDirPath()}/volume/${id}.json`)
   return true
 }
